+++ oasis/llvm
@@ -41,8 +41,8 @@ Library bap_llvm
                  llvm_binary.h,
                  llvm_binary_34.c,
                  llvm_binary_34.h,
-                 llvm_binary_38.c,
-                 llvm_binary_38.h,
+                 llvm_binary_38_40.c,
+                 llvm_binary_38_40.h,
                  llvm_error_or.c,
                  llvm_error_or.h,
                  llvm_loader.c,
